% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dividend_yield.R
\name{dividend_yield}
\alias{dividend_yield}
\title{Dividend Yield}
\usage{
dividend_yield(dividend_per_share, share_price)
}
\arguments{
\item{dividend_per_share:}{The annual dividend paid per share.}

\item{share_price:}{The current market price of the share.}
}
\description{
Calculates the dividend yield based on the given dividend per share and share price.
}
\examples{
dividend_yeild()
}
\keyword{Market}
\keyword{Stock}
