% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMA.R
\name{EMA}
\alias{EMA}
\title{Exponential Moving Average (EMA)}
\usage{
EMA(prices, n)
}
\arguments{
\item{prices:}{A vector of prices.}

\item{n:}{The number of periods to include in the EMA calculation.}
}
\value{
The function returns a vector containing the calculated EMA values for the entire data series.
EMA()
}
\description{
This function allows you to calculate the Smoothing Moving Average.
}
\keyword{Market}
\keyword{Stock}
