% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PE_ratio.R
\name{PE_ratio}
\alias{PE_ratio}
\title{Price-Earnings Ratio (P/E Ratio)}
\usage{
PE_ratio(share_price, EPS)
}
\arguments{
\item{share_price:}{The current market price of the stock.}

\item{EPS:}{The Earnings Per Share of the company.}
}
\description{
The Price-Earnings Ratio (P/E Ratio) is a valuation ratio that compares a company's current share price to its earnings per share (EPS).
}
\examples{
PE_ratio()
}
\keyword{Market}
\keyword{Stock}
