% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEG_ratio.R
\name{PEG_ratio}
\alias{PEG_ratio}
\title{Price-to-Earnings Growth (PEG) Ratio}
\usage{
PEG_ratio(share_price, EPS, Earnings_Growth_Rate)
}
\arguments{
\item{share_price:}{The current market price of the stock.}

\item{EPS:}{The Earnings Per Share of the company.}

\item{Earnings_Growth_Rate:}{The expected annual earnings growth rate.}
}
\description{
The PEG ratio is calculated by taking the P/E ratio of a company and dividing it by the year-over-year growth rate of its earnings as an estimate going forward.
}
\examples{
PEG_ratio()
}
\keyword{Market}
\keyword{Stock}
