% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMA.R
\name{SMA}
\alias{SMA}
\title{Smoothing Moving Average}
\usage{
SMA(prices, n)
}
\arguments{
\item{prices:}{A vector of prices.}

\item{n:}{The number of periods to include in the SMA calculation.}
}
\description{
This function allows you to calculate the Smoothing Moving Average.
}
\examples{
SMA()
}
\keyword{Market}
\keyword{Stock}
